LayoutTests/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/parsing/masonry-shorthand-valid-expected.txt


FAIL masonry followed by !important assert_equals: expected (string) "\"test\" max-content row normal" but got (undefined) undefined
FAIL e.style['masonry'] = "\"a\" calc(10px) column-reverse normal" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['masonry'] = "minmax(calc(30% + 40vw), 10px)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL masonry: minmax(10px, 20px) row should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL masonry: 1px 2px should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL masonry: "a" 10px reverse should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL masonry: "a b" 10px 20px row normal should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL masonry: "a b c" 10% 20% 30% row-reverse should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL masonry: repeat(5, auto) row reverse should be valid. assert_not_equals: property should be set got disallowed value ""
FAIL e.style['masonry'] = "none" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['masonry'] = "none" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "none" but got ""
FAIL e.style['masonry'] = "none" should set masonry-direction assert_equals: masonry-direction should be canonical expected (string) "column" but got (undefined) undefined
FAIL e.style['masonry'] = "none" should set masonry-fill assert_equals: masonry-fill should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['masonry'] = "none" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['masonry'] = "10px reverse" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['masonry'] = "10px reverse" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "10px" but got ""
FAIL e.style['masonry'] = "10px reverse" should set masonry-direction assert_equals: masonry-direction should be canonical expected (string) "column" but got (undefined) undefined
FAIL e.style['masonry'] = "10px reverse" should set masonry-fill assert_equals: masonry-fill should be canonical expected (string) "reverse" but got (undefined) undefined
FAIL e.style['masonry'] = "10px reverse" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['masonry'] = "\"b a\" 20% 40% column normal" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "\"b a\"" but got ""
FAIL e.style['masonry'] = "\"b a\" 20% 40% column normal" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "20% 40%" but got ""
FAIL e.style['masonry'] = "\"b a\" 20% 40% column normal" should set masonry-direction assert_equals: masonry-direction should be canonical expected (string) "column" but got (undefined) undefined
FAIL e.style['masonry'] = "\"b a\" 20% 40% column normal" should set masonry-fill assert_equals: masonry-fill should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['masonry'] = "\"b a\" 20% 40% column normal" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['masonry'] = "\"b b a\" 1fr 2fr 3fr row" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "\"b\" \"b\" \"a\"" but got ""
FAIL e.style['masonry'] = "\"b b a\" 1fr 2fr 3fr row" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "1fr 2fr 3fr" but got ""
FAIL e.style['masonry'] = "\"b b a\" 1fr 2fr 3fr row" should set masonry-direction assert_equals: masonry-direction should be canonical expected (string) "row" but got (undefined) undefined
FAIL e.style['masonry'] = "\"b b a\" 1fr 2fr 3fr row" should set masonry-fill assert_equals: masonry-fill should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['masonry'] = "\"b b a\" 1fr 2fr 3fr row" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['masonry'] = "repeat(2, auto) row-reverse" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['masonry'] = "repeat(2, auto) row-reverse" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "repeat(2, auto)" but got ""
FAIL e.style['masonry'] = "repeat(2, auto) row-reverse" should set masonry-direction assert_equals: masonry-direction should be canonical expected (string) "row-reverse" but got (undefined) undefined
FAIL e.style['masonry'] = "repeat(2, auto) row-reverse" should set masonry-fill assert_equals: masonry-fill should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['masonry'] = "repeat(2, auto) row-reverse" should not set unrelated longhands assert_true: expected true got false
